ファイルの分割化、次はLinkOverlayコンポーネントに取り掛かる
