+++ components/navigation/src/index.ts
@@ -1,1 +1,2 @@
 export * from './link'
+export * from './linkBox'
